+++ test/solvers/test_particle_swarm.jl
@@ -67,4 +67,8 @@ using Random
         s = particle_swarm(M, f, swarm)
         @test s ≈ 0.0
     end
+    @testset "Specific Stopping criteria" begin
+        sc = StopWhenSwarmVelocityLess(1.0)
+        @test startswith(repr(sc), "StopWhenSwarmVelocityLess")
+    end
 end